+++ contract/lib/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-alpha.13](https://github.com/KYVENetwork/kyve/compare/@kyve/contract-lib@0.1.0-alpha.12...@kyve/contract-lib@0.1.0-alpha.13) (2021-07-12)
+
+
+### Features
+
+* added governance contract to contract-lib ([cf9a6d0](https://github.com/KYVENetwork/kyve/commit/cf9a6d033c19c6e1330551fcadbb135778d7a5d5))
+
+
+
+
+
+# [0.1.0-alpha.12](https://github.com/KYVENetwork/kyve/compare/@kyve/contract-lib@0.1.0-alpha.11...@kyve/contract-lib@0.1.0-alpha.12) (2021-07-09)
+
+
+### Bug Fixes
+
+* added missing cross-fetch dependency ([adb27ad](https://github.com/KYVENetwork/kyve/commit/adb27ad6dbbb67a1ca77eeeb4668cae28e0983fb))
+
+
+### Features
+
+* added cache to get pool state ([5b88a82](https://github.com/KYVENetwork/kyve/commit/5b88a829c895dc9b03cdacfdf3711f8ca183a1bc))
+
+
+
+
+
 # [0.1.0-alpha.11](https://github.com/KYVENetwork/kyve/compare/@kyve/contract-lib@0.1.0-alpha.10...@kyve/contract-lib@0.1.0-alpha.11) (2021-07-07)
 
 **Note:** Version bump only for package @kyve/contract-lib