+++ integrations/smartweave/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.0.0-alpha.14](https://github.com/KYVENetwork/smartweave/compare/@kyve/smartweave@0.0.0-alpha.13...@kyve/smartweave@0.0.0-alpha.14) (2021-07-12)
+
+**Note:** Version bump only for package @kyve/smartweave
+
+
+
+
+
+# [0.0.0-alpha.13](https://github.com/KYVENetwork/smartweave/compare/@kyve/smartweave@0.0.0-alpha.12...@kyve/smartweave@0.0.0-alpha.13) (2021-07-09)
+
+**Note:** Version bump only for package @kyve/smartweave
+
+
+
+
+
 # [0.0.0-alpha.12](https://github.com/KYVENetwork/smartweave/compare/@kyve/smartweave@0.0.0-alpha.11...@kyve/smartweave@0.0.0-alpha.12) (2021-07-07)
 
 **Note:** Version bump only for package @kyve/smartweave